requirements.txt
python-dotenv
odfpy
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
groq==0.13.0
streamlit==1.40.2